Add put, post, and delete (#2)
* Adding Person object with JPA - H2

* Adding endpoint to retrieve specific data with additional custom error handling

* Addind custom error handler to return JSON response when the error occurs

* Addind insert endpoint with JPA column validation

* Addind delete endpoint for specific person

* Addind update endpoint for specific person

+++ src/main/java/com/mjubilee/communitiessapims/repository/PersonRepository.java
@@ -1,11 +1,9 @@
 package com.mjubilee.communitiessapims.repository;
 
-import java.math.BigDecimal;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.mjubilee.communitiessapims.model.Person;
 
-public interface PersonRepository extends JpaRepository<Person, BigDecimal> {
+public interface PersonRepository extends JpaRepository<Person, Long> {
+
 
 }
